requirements.txt
Flask==0.10.1
Flask-Security==1.7.5
GitPython==1.0.1
nose==1.3.7
psycopg2==2.6.1
SQLAlchemy==1.0.9
Flask-SQLAlchemy==2.1
Flask-Mail==0.9.1
SQLAlchemy-Searchable==0.9.3